Hide fields that are empty in secrets configuration
Updated the existing helper to check for empty object, string,
null or undefined
