--- ui/app/helpers/is-empty-object.js
@@ -1,5 +1,0 @@
-import { helper } from '@ember/component/helper';
-
-export default helper(function isEmptyObject([object] /*, hash*/) {
-  return Object.keys(object).length === 0;
-});
