+++ ui/app/templates/vault/cluster/secrets/backend/configuration.hbs
@@ -20,9 +20,9 @@
 <div class="box is-fullwidth is-sideless is-paddingless is-marginless">
   {{#each model.attrs as |attr|}}
     {{#if (eq attr.type "object")}}
-      <InfoTableRow @alwaysRender={{true}} @label={{or attr.options.label (to-label attr.name)}} @value={{stringify (get model attr.name)}} />
+      <InfoTableRow @alwaysRender={{not (is-empty-value (get model attr.name))}} @label={{or attr.options.label (to-label attr.name)}} @value={{stringify (get model attr.name)}} />
     {{else}}
-      <InfoTableRow @alwaysRender={{not-eq attr.name "options.version"}} @label={{or attr.options.label (to-label attr.name)}} @value={{get model attr.name}} />
+      <InfoTableRow @alwaysRender={{and (not (is-empty-value (get model attr.name))) (not-eq attr.name "options.version")}} @label={{or attr.options.label (to-label attr.name)}} @value={{get model attr.name}} />
     {{/if}}
   {{/each}}
 </div>
